230822 엔티티 생성(2)

+++ src/main/java/com/dukbab/repository/CartItemRepository.java
@@ -3,5 +3,5 @@ package com.dukbab.repository;
 import com.dukbab.domain.CartItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CartItemRepository extends JpaRepository<CartItem, Long> {
+public interface CartItemRepository extends JpaRepository<CartItem, Integer> {
 }
